--- a/tensordict/_td.py
+++ b/tensordict/_td.py
@@ -2448,6 +2448,22 @@ class TensorDict(TensorDictBase):
     ) -> Iterator[tuple[str, CompatibleType]]:
         if not include_nested and not leaves_only:
             return self._tensordict.items()
+        elif include_nested and leaves_only:
+            is_leaf = _default_is_leaf if is_leaf is None else is_leaf
+            def fast_iter():
+                for k, val in self._tensordict.items():
+                    if not is_leaf(val.__class__):
+                        yield from (
+                            ((k, *((_key,) if isinstance(_key, str) else _key)), _val)
+                            for _key, _val in val.items(
+                                include_nested=include_nested,
+                                leaves_only=leaves_only,
+                                is_leaf=is_leaf,
+                            )
+                        )
+                    else:
+                        yield k, val
+            return fast_iter()
         else:
             return super().items(
                 include_nested=include_nested, leaves_only=leaves_only, is_leaf=is_leaf