--- a/tensordict/base.py
+++ b/tensordict/base.py
@@ -6212,17 +6212,18 @@ To temporarily permute a tensordict you can still user permute() as a context ma
         """
         if is_leaf is None:
             is_leaf = _is_leaf_nontensor
-        all_leaves = list(
-            self.keys(include_nested=True, leaves_only=True, is_leaf=is_leaf)
+        all_leaves_items = list(
+            self.items(include_nested=True, leaves_only=True, is_leaf=is_leaf)
         )
         all_leaves_flat = [
-            separator.join(key) if isinstance(key, tuple) else key for key in all_leaves
+            separator.join(key) if isinstance(key, tuple) else key for key, _ in all_leaves_items
         ]
-        if len(set(all_leaves_flat)) < len(set(all_leaves)):
+
+        if len(set(all_leaves_flat)) < len(all_leaves_flat):
             # find duplicates
             seen = set()
             conflicts = []
-            for leaf, leaf_flat in zip(all_leaves, all_leaves_flat):
+            for (leaf, _), leaf_flat in zip(all_leaves_items, all_leaves_flat):
                 if leaf_flat in seen:
                     conflicts.append(leaf)
                 else:
@@ -6233,7 +6234,7 @@ To temporarily permute a tensordict you can still user permute() as a context ma
         if inplace:
             # we will need to remove the empty tensordicts later on
             root_keys = set(self.keys())
-            for leaf, leaf_flat in zip(all_leaves, all_leaves_flat):
+            for (leaf, _), leaf_flat in zip(all_leaves_items, all_leaves_flat):
                 self.rename_key_(leaf, leaf_flat)
                 if isinstance(leaf, str):
                     root_keys.discard(leaf)
@@ -6241,14 +6242,19 @@ To temporarily permute a tensordict you can still user permute() as a context ma
             return self
         else:
             result = self.empty()
-            for leaf, leaf_flat in zip(all_leaves, all_leaves_flat):
-                result._set_str(
-                    leaf_flat,
-                    self.get(leaf),
-                    validated=True,
-                    inplace=False,
-                    non_blocking=False,
-                )
+            if hasattr(result, "_set_dict"):
+                result._set_dict({
+                    k: v for (_, v), k in zip(all_leaves_items, all_leaves_flat)
+                }, validated=True)
+            else:
+                for (leaf, _), leaf_flat in zip(all_leaves_items, all_leaves_flat):
+                    result._set_str(
+                        leaf_flat,
+                        self.get(leaf),
+                        validated=True,
+                        inplace=False,
+                        non_blocking=False,
+                    )
             # Uncomment if you want key operations to propagate the shared status
             # self._maybe_set_shared_attributes(result)
             if result._is_shared or result._is_memmap: