--- a/tensordict/utils.py
+++ b/tensordict/utils.py
@@ -1016,12 +1016,12 @@ def _unfold_sequence(seq):
 
 
 def _make_cache_key(args, kwargs):
-    """Creats a key for the cache such that memory footprint is minimized."""
+    """Creates a key for the cache such that memory footprint is minimized."""
     # Fast path for the common args
     if not args and not kwargs:
-        key = ((), ())
+        return ((), ())
     elif not kwargs and len(args) == 1 and type(args[0]) is str:
-        key = (args, ())
+        return (args, ())
     else:
         return (
             tuple(_unfold_sequence(args)),