@vmoens 2024-04-18
For some reason the benchmarks don't appear in this PR but you can check them [here](https://github.com/pytorch/tensordict/actions/runs/8733935072/job/23963570632?pr=728)
they look pretty good! Thanks so much!